Initial implementation of on the fly re-encryption of backup.
