+++ entity-store/build.gradle
@@ -4,6 +4,7 @@ dependencies {
     implementation project(':xodus-utils')
     implementation project(':xodus-environment')
     implementation project(':xodus-compress')
+    implementation 'commons-io:commons-io:2.11.0'
     testImplementation 'io.github.classgraph:classgraph:4.8.90'
     testImplementation project(':xodus-utils').sourceSets.test.output
 }
